backend/requirements.txt
fastapi
uvicorn[standard]
openai
google-generativeai
python-dotenv
pydantic